# back/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Profesto",
        "description": "Back End sistema control ingreso y egreso personal",
        "version": "1.0.0"
    },
    "paths": {
        "/api/auth/register": {
            "post": {
                "tags": [
                    "AuthController"
                ],
                "summary": "Registro de usuario",
                "operationId": "register",
                "parameters": [
                    {
                        "name": "Register",
                        "in": "query",
                        "schema": {
                            "$ref": "#/components/schemas/RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Usuario creado",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Solicitud no válida"
                    },
                    "404": {
                        "description": "No encontrado"
                    },
                    "422": {
                        "description": "Error validación"
                    }
                }
            }
        },
        "/api/auth/login": {
            "post": {
                "tags": [
                    "AuthController"
                ],
                "summary": "Login de usuario",
                "operationId": "login",
                "parameters": [
                    {
                        "name": "Login",
                        "in": "query",
                        "schema": {
                            "$ref": "#/components/schemas/LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Usuario creado",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Solicitud no válida"
                    },
                    "401": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "No encontrado"
                    },
                    "422": {
                        "description": "Error validación"
                    }
                }
            }
        },
        "/api/auth/logout": {
            "get": {
                "tags": [
                    "AuthController"
                ],
                "summary": "Logout de usuario",
                "operationId": "logout",
                "responses": {
                    "201": {
                        "description": "Usuario creado",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Solicitud no válida"
                    },
                    "401": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "No encontrado"
                    },
                    "422": {
                        "description": "Error validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/user": {
            "get": {
                "tags": [
                    "AuthController"
                ],
                "summary": "Datos del usuario",
                "operationId": "user",
                "responses": {
                    "200": {
                        "description": "Usuario",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Solicitud no válida"
                    },
                    "401": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "No encontrado"
                    },
                    "422": {
                        "description": "Error validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/empleado": {
            "get": {
                "tags": [
                    "EmpleadoController"
                ],
                "summary": "Muestra todos los empleados",
                "operationId": "empleadoIndex",
                "responses": {
                    "200": {
                        "description": "Empleados",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Solicitud no válida"
                    },
                    "404": {
                        "description": "No encontrado"
                    },
                    "422": {
                        "description": "Error validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "EmpleadoController"
                ],
                "summary": "Modificar un empleado",
                "operationId": "empleadoUpdate",
                "parameters": [
                    {
                        "name": "EmpleadoUpdate",
                        "in": "query",
                        "schema": {
                            "$ref": "#/components/schemas/EmpleadoUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Empleado modificado",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Solicitud no válida"
                    },
                    "404": {
                        "description": "No encontrado"
                    },
                    "422": {
                        "description": "Error validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "EmpleadoController"
                ],
                "summary": "Cargar un empleado",
                "operationId": "empleadoStore",
                "parameters": [
                    {
                        "name": "Register",
                        "in": "query",
                        "schema": {
                            "$ref": "#/components/schemas/EmpleadoStoreRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Empleado creado",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Solicitud no válida"
                    },
                    "404": {
                        "description": "No encontrado"
                    },
                    "422": {
                        "description": "Error validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/empleado/{id}": {
            "get": {
                "tags": [
                    "EmpleadoController"
                ],
                "summary": "Buscar un empleado",
                "operationId": "empleadoShow",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id del empleado",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Empleado",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Solicitud no válida"
                    },
                    "404": {
                        "description": "No encontrado"
                    },
                    "422": {
                        "description": "Error validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/empleado/showPorUsuario/{idUser}": {
            "get": {
                "tags": [
                    "EmpleadoController"
                ],
                "summary": "Empleados de un usuario en particular",
                "operationId": "empleadosPorUsuario",
                "parameters": [
                    {
                        "name": "idUser",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Usuario creado",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Solicitud no válida"
                    },
                    "401": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "No encontrado"
                    },
                    "422": {
                        "description": "Error validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Empleado": {
                "title": "Empleado",
                "description": "Representación del empleado",
                "properties": {
                    "id": {
                        "description": "Id de empleado",
                        "type": "integer"
                    },
                    "apellido": {
                        "description": "Apellido del empleado",
                        "type": "string"
                    },
                    "nombre": {
                        "description": "nombre del empleado",
                        "type": "string"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    }
                },
                "type": "object"
            },
            "EmpleadoStoreRequest": {
                "title": "EmpleadoStoreRequest",
                "description": "Empleados Store Request",
                "properties": {
                    "user_id": {
                        "description": "Id del usuario",
                        "type": "integer"
                    },
                    "nombre": {
                        "description": "nombre del empleado",
                        "type": "string"
                    },
                    "apellido": {
                        "description": "Apellido del empleado",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "EmpleadoUpdateRequest": {
                "title": "EmpleadoUpdateRequest",
                "description": "Empleados Update Request",
                "properties": {
                    "id": {
                        "description": "Id del empleado",
                        "type": "integer"
                    },
                    "nombre": {
                        "description": "nombre del empleado",
                        "type": "string"
                    },
                    "apellido": {
                        "description": "Apellido del empleado",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "LoginRequest": {
                "title": "LoginRequest",
                "description": "Login Request",
                "properties": {
                    "email": {
                        "description": "Email del usuario",
                        "type": "string",
                        "format": "email"
                    },
                    "password": {
                        "description": "Contraseña del usuario",
                        "type": "string",
                        "format": "password"
                    },
                    "remember_me": {
                        "description": "Recordar conexión del usuario",
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
            "RegisterRequest": {
                "title": "RegisterRequest",
                "description": "Register Request",
                "properties": {
                    "email": {
                        "description": "Email del usuario",
                        "type": "string",
                        "format": "email"
                    },
                    "password": {
                        "description": "Contraseña del usuario",
                        "type": "string",
                        "format": "password"
                    }
                },
                "type": "object"
            },
            "User": {
                "title": "User",
                "description": "User representation",
                "properties": {
                    "id": {
                        "description": "The user id",
                        "type": "integer"
                    },
                    "name": {
                        "description": "Nombre del usuario",
                        "type": "string"
                    },
                    "email": {
                        "description": "Email del usuario",
                        "type": "string",
                        "format": "email"
                    },
                    "email_verified_at": {
                        "description": "Cuando el usuario verifica su email",
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "active": {
                        "description": "Si usuario se encuentra activo",
                        "type": "boolean"
                    },
                    "created_at": {
                        "description": "Cuando el usuario fue creado",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Cuando el usuario fue modificado la última vez",
                        "type": "string",
                        "format": "date-time"
                    },
                    "deleted_at": {
                        "description": "When user was last updated",
                        "type": "string",
                        "format": "date-time"
                    },
                    "email_alternativo": {
                        "description": "Email alternativo del usuario",
                        "type": "string",
                        "format": "email"
                    },
                    "": {
                        "$ref": "#/components/schemas/Empleado"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}